refactor log_setting
