+++ src/otaclient/app/configs.py
@@ -97,6 +97,7 @@ class BaseConfig(_InternalSettings):
         "otaclient": INFO,
         "otaclient_api": INFO,
         "otaclient_common": INFO,
+        "otaproxy": INFO,
     }
     LOG_FORMAT = (
         "[%(asctime)s][%(levelname)s]-%(name)s:%(funcName)s:%(lineno)d,%(message)s"
