+++ src/otaclient/app/main.py
@@ -29,7 +29,7 @@ import ota_metadata.legacy  # noqa: F401
 from otaclient import __version__
 from otaclient.app.configs import config as cfg
 from otaclient.app.configs import ecu_info, server_cfg
-from otaclient.app.log_setting import configure_logging
+from otaclient.log_setting import configure_logging
 from otaclient.app.ota_client_stub import OTAClientServiceStub
 from otaclient_api.v2 import otaclient_v2_pb2_grpc as v2_grpc
 from otaclient_api.v2.api_stub import OtaClientServiceV2
